+++ tsfile/src/test/java/org/apache/iotdb/tsfile/read/TsFileSequenceReaderTest.java
@@ -24,12 +24,9 @@ import org.apache.iotdb.tsfile.file.MetaMarker;
 import org.apache.iotdb.tsfile.file.header.ChunkGroupHeader;
 import org.apache.iotdb.tsfile.file.header.ChunkHeader;
 import org.apache.iotdb.tsfile.file.header.PageHeader;
-import org.apache.iotdb.tsfile.file.metadata.AlignedChunkMetadata;
 import org.apache.iotdb.tsfile.file.metadata.ChunkMetadata;
-import org.apache.iotdb.tsfile.file.metadata.IChunkMetadata;
 import org.apache.iotdb.tsfile.utils.FileGenerator;
 import org.apache.iotdb.tsfile.utils.Pair;
-import org.apache.iotdb.tsfile.utils.TsFileGeneratorForTest;
 
 import org.junit.After;
 import org.junit.Assert;
@@ -127,42 +124,5 @@ public class TsFileSequenceReaderTest {
     // test for non-exist device "d3"
     Assert.assertTrue(reader.readChunkMetadataInDevice("d3").isEmpty());
     reader.close();
-  }
-
-  @Test
-  public void testGetAlignedChunkMetadata() throws IOException {
-    // generate aligned timeseries "d1.s1","d1.s2","d1.s3","d1.s4" and nonAligned timeseries
-    // "d2.s1","d2.s2","d2.s3"
-    TsFileGeneratorForTest.generateAlignedTsFile(10, 100, 30);
-    String filePath = TsFileGeneratorForTest.alignedOutputDataFile;
-    try (TsFileSequenceReader reader = new TsFileSequenceReader(filePath)) {
-      // query for non-exist device
-      try {
-        reader.getAlignedChunkMetadata("d3");
-      } catch (IOException e) {
-        Assert.assertEquals("Device {d3} is not in tsFileMetaData", e.getMessage());
-      }
-
-      // query for non-aligned device
-      try {
-        reader.getAlignedChunkMetadata("d2");
-      } catch (IOException e) {
-        Assert.assertEquals("Timeseries of device {d2} are not aligned", e.getMessage());
-      }
-
-      String[] expected = new String[] {"s1", "s2", "s3", "s4"};
-
-      List<AlignedChunkMetadata> chunkMetadataList = reader.getAlignedChunkMetadata("d1");
-      AlignedChunkMetadata alignedChunkMetadata = chunkMetadataList.get(0);
-      Assert.assertEquals("", alignedChunkMetadata.getTimeChunkMetadata().getMeasurementUid());
-      int i = 0;
-      for (IChunkMetadata chunkMetadata : alignedChunkMetadata.getValueChunkMetadataList()) {
-        Assert.assertEquals(expected[i], chunkMetadata.getMeasurementUid());
-        i++;
-      }
-
-      Assert.assertEquals(expected.length, i);
-    }
-    TsFileGeneratorForTest.closeAlignedTsFile();
   }
 }
